+++ src/components/StatusBadge/StatusBadgeTypes.tsx
@@ -1,0 +1,5 @@
+export enum StatusBadgeTypes {
+  Low = "low",
+  Medium = "medium",
+  High = "high",
+}