+++ CHANGELOG.md
@@ -21,6 +21,8 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 - Composes the DS `TemplateAppContainer` component from Chakra's `Box` component for page layouts through props.
 - Composes the DS `Link` component with Chakra component and fixes the `Button` type.
 - Exports Chakra's `Grid` and `GridItem` components.
+- Composes the DS `Label` component with Chakra component and updates the font size.
+- Composes the DS `StatusBadge` component from Chakra's component.
 - Composes the DS `Icon` component with Chakra's `Icon` component. Added new props for `Icon` sizing, alignment, name, and rotation.
 
 ### Changes